Statistical tests for the differences between standardized accuracies (staccuracies)

Because the distribution of staccuracies is uncertain (and indeed, different staccuracies likely have different distributions), bootstrapping is used to empirically estimate the distributions and calculate the p-values.

Usage

sa_diff(
  actual,
  preds,
  ...,
  na.rm = FALSE,
  sa = NULL,
  pct = c(0.01, 0.02, 0.03, 0.04, 0.05),
  boot_alpha = 0.05,
  boot_it = 1000,
  seed = 0
)

Arguments

actual

numeric vector. The actual (true) labels.

preds

named list of at least two numeric vectors. Each element is a vector of the same length as actual with predictions for each row corresponding to each element of actual. The names of the list elements will be used for the columns of the result.

...

not used. Forces explicit naming of subsequent arguments.

na.rm

See documentation for staccuracy()

sa

list of functions. Each element is the unquoted name of a staccuracy function. If an element is named, the name will be displayed as the value of the sa column of the result. Otherwise, the function name will be displayed. If NULL (default), staccuracy functions will be automatically selected based on the datatypes of actual and preds.

pct

numeric with values from (0, 1). The percentage values on which the difference in staccuracies will be tested.

boot_alpha

numeric(1) from 0 to 1. Alpha for percentile-based confidence interval range for the bootstrapped means; the bootstrap confidence intervals will be the lowest and highest (1 - 0.05) / 2 percentiles. For example, if boot_alpha = 0.05 (default), the intervals will be at the 2.5 and 97.5 percentiles.

boot_it

positive integer(1). The number of bootstrap iterations.

seed

integer(1). Random seed for the bootstrap sampling. Supply this between runs to assure identical results.

Value

tibble. Columns are sa (name of staccuracy measure), then a column for each named element in preds (staccuracy for that prediction), then a column for each element of pct (p-value; percentage of iterations where the difference between staccuracy in the pair of preds is equal or greater than the pct value). E.g., for the default pct = c(0.01, 0.02, 0.03, 0.04, 0.05), these columns would be p_1, p_2, p_3, p_4, and p_5. When there are more than two predictions, then each row will compare only two at a time.

Examples

lm_attitude_all <- lm(rating ~ ., data = attitude)
lm_attitude__c <- lm(rating ~ . - complaints, data = attitude)

sa_diff(
  attitude$rating,
  list(
    all = predict(lm_attitude_all),
    madv = predict(lm_attitude__c)
  ),
  boot_it = 10
)
#> # A tibble: 2 × 15
#>   staccuracy        all_lo all_mn all_hi madv_lo madv_mn madv_hi diff_lo diff_mn
#>   <chr>              <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 Staccuracy WinMA…  0.672  0.719  0.776   0.586   0.635   0.692  0.0440  0.0840
#> 2 Staccuarcy WinRM…  0.684  0.737  0.781   0.616   0.670   0.723  0.0335  0.0666
#> # ℹ 6 more variables: diff_hi <dbl>, p_01 <dbl>, p_02 <dbl>, p_03 <dbl>,
#> #   p_04 <dbl>, p_05 <dbl>